Split classes in to their own files, facilitate running in dev
This commit splits the single lambda function file into a file per
class.

It also adds a method for helping to test in dev.
