+++ README.md
@@ -1,0 +1,3 @@
+### Using this locally
+Use the `#mock_http_for_lambda` method. Make sure your AWS credentials
+are good.
